+++ src/components/page/Magazine/MagazinesList/index.ts
@@ -1,0 +1,1 @@
+export { MagazinesList } from './MagazinesList';
